+++ src/sbs/sbs.cpp
@@ -3514,12 +3514,12 @@ void SBS::ListVisibleMeshes()
 
 	for (size_t i = 0; i < dynamic_meshes.size(); i++)
 	{
-		for (size_t j = 0; j < dynamic_meshes[i]->GetMeshCount(); j++)
+		for (int j = 0; j < dynamic_meshes[i]->GetMeshCount(); j++)
 		{
-			if (camera->IsDynamicMeshVisible(dynamic_meshes[i], (int)j) == true)
+			if (camera->IsDynamicMeshVisible(dynamic_meshes[i], j) == true)
 			{
-				submeshes = dynamic_meshes[i]->GetSubMeshCount((int)j);
-				Report(dynamic_meshes[i]->GetMeshName((int)j) + "\t-\t" + ToString(submeshes));
+				submeshes = dynamic_meshes[i]->GetSubMeshCount(j);
+				Report(dynamic_meshes[i]->GetMeshName(j) + "\t-\t" + ToString(submeshes));
 				count++;
 				total += submeshes;
 			}
